requirements.txt
Django==4.1.1
djangorestframework==3.14.0
django-bootstrap5==22.1
Pillow==9.3.0